action: add MODEL_FIXUP_RELSEQ action_type
This type will be used as a special model-checker action for fixing up
release sequences. Each action corresponds to "finalizing" one release
sequence: either force a particular write to break the sequence or else
allow the synchronization to occur.

Currently, this action_type won't do anything, as it's not hooked up in
ModelChecker.